requirements.txt
pandas
numpy
ta
ccxt
tensorflow
python-dotenv
requests
sqlalchemy
pyinstaller
scikit-learn
schedule
yfinance
streamlit
plotly
joblib
